Short-circuit and log on bogus directions
Former-commit-id: 7416b817d820f5d109bb280725cfd6a1372c5b75

+++ Intersect.Server/Classes/Entities/Player.cs
@@ -4083,7 +4083,7 @@ namespace Intersect.Server.Entities
         }
 
 
-        public override void Move(byte moveDir, Client client, bool doNotUpdate = false, bool correction = false)
+        public override void Move(int moveDir, Client client, bool doNotUpdate = false, bool correction = false)
         {
             var oldMap = MapId;
             client = Client;
